+++ experimental/CHANGELOG.md
@@ -8,6 +8,8 @@ All notable changes to experimental packages in this project will be documented 
 
 ### :rocket: (Enhancement)
 
+* feat(exporters): update proto version and use otlp-transformer #2929 @pichlermarc
+
 ### :bug: (Bug Fix)
 
 ### :books: (Refine Doc)